[MXNet] detection and semantic segmentation (#402)
